refactor(CBS-256567:LN-58416:LN-58675): enhance approval strategy with comprehensive implementation
- Complete AutoApprovalStrategy and ManualApprovalStrategy implementations with business logic
- Enhance ApprovalStrategyFactory with intelligent strategy selection based on facility criteria
- Update ApproveFacilityCommand to support new approval workflow and sanction verification
- Refactor REST controllers and DTOs to integrate with approval strategy pattern
- Add comprehensive internationalization support for approval process messages
- Update persistence entities to support approval workflow states
- Enhance domain events and services for approval strategy integration
- Update database changelog master to include new sanction details schema
- Improve command handlers across facility lifecycle operations

+++ adapters/driving/rest/src/main/java/ir/dotin/loan/trade/adapters/driving/rest/command/dto/ApproveFacilityRequest.java
@@ -14,4 +14,4 @@ public record ApproveFacilityRequest(
                 @NotNull
                 UUID uid,
         @Schema(description = "نسخه عملیات", example = "1", requiredMode = Schema.RequiredMode.REQUIRED) @NotNull
-                Integer version) {}
+                Long version) {}
